requirements.txt
streamlit
langchain-google-genai
langchain-community
faiss-cpu
PyPDF2
sentence-transformers
google-genai
python-dotenv